chore: update dataset file structure
